Added some of the possible ugliestly namedely exceptionsly that addalyse can judiciously raise.
Am pondering on adding the fantastic AddalyseUserNotInSolrButIWantedToRemakeTheirProfileError for the case when we try to do the
update case when the user is not already in Solr. However current behaviour is to silently do a regular add in this case instead
of violently exploding using some obscure exception with a very very bad name.